Branches with staging repo (#1)
After many tries, this works. It checks out the repository on the current branch, then it compiles the index.html file, and places it into a folder with the name of the branch in the staging repository which is also checked out. Finally, the staging repository is pushed. This works using a deployment key configured on the deploying (secret ssh deployment key) and staging (public ssh key) repositories.